Move PendingSnapshot class into PassiveRole.
